requirements.txt
# Core
faiss-cpu
langchain>=0.2.0
langchain-community
transformers
torch
sentence-transformers
pandas
numpy
jq
# Embeddings
huggingface-hub

# Mistral (SDK officiel ou wrapper)
mistralai

# API
fastapi
uvicorn

# Utils
python-dotenv
requests

# Tests
pytest
ragas
datasets
evaluate
bs4
pydantic